util: Separate which is platform dependent
For instance, RIOT has not file system, so there is no reason to have
code to deal with it.
