+++ src/shared/Makefile
@@ -22,7 +22,8 @@ ifeq (y,$(SOL_PLATFORM_LINUX))
 obj-libshared-m += \
     sol-conffile.o \
     sol-file-reader.o \
-    sol-network-linux.o
+    sol-network-linux.o \
+    sol-util-linux.o
 obj-libshared-m-extra-cflags += $(GLIB_CFLAGS)
 obj-libshared-m-extra-ldflags += $(GLIB_LDFLAGS)
 endif
